requirements.txt
# Web framework
Flask>=2.0,<3.0
jinja2>=3.0,<4.0

# OCR & Image processing
pytesseract>=0.3.10,<0.4
Pillow>=9.0,<10.0

# Data handling & Excel export
numpy==1.26.4
pandas>=1.3,<2.0
matplotlib>=3.5.0,<4.0
openpyxl>=3.0,<4.0

# PDF generation
weasyprint==57.2

# Email sending
yagmail

# Environment variable management
python-dotenv>=0.21.0,<0.22
# Trigger update

gunicorn>=20.0.0

pydyf==0.8.0

# Ensure charts render properly
matplotlib>=3.5.0,<4.0

# Dev tools
pytest>=7.0.0
dash==3.0.4
dash-bootstrap-components==2.0.3

black>=23.3.0

fpdf2==2.7.8

# Plot rendering
kaleido>=0.2.1

flask-jwt-extended>=4.4.4

flask-limiter>=3.5.0

sentry-sdk>=1.40.0

PyYAML>=6.0

pydantic>=1.10,<2.0

flask-smorest>=0.42.0
